+++ DearPyGui/src/core/AppItems/basic/mvCombo.cpp
@@ -64,7 +64,12 @@ namespace Marvel {
 				if (ImGui::Selectable((name).c_str(), is_selected))
 				{
 					if (m_enabled) { *m_value = name; }
-					mvApp::GetApp()->getCallbackRegistry().addCallback(m_callback, m_uuid, nullptr, m_user_data);
+	
+					auto value = *m_value;
+					mvApp::GetApp()->getCallbackRegistry().submitCallback([=]() {
+						mvApp::GetApp()->getCallbackRegistry().addCallback(getCallback(false), m_uuid, ToPyString(value), m_user_data);
+						});
+
 
 				}