+++ DearPyGui/src/core/AppItems/basic/mvDragSingle.cpp
@@ -113,7 +113,12 @@ namespace Marvel {
         if (!m_enabled) m_disabled_value = *m_value;
 
         if (ImGui::DragFloat(m_label.c_str(), m_enabled ? m_value.get() : &m_disabled_value, m_speed, m_min, m_max, m_format.c_str(), m_flags))
-            mvApp::GetApp()->getCallbackRegistry().addCallback(getCallback(false), m_uuid, nullptr, m_user_data);
+        {
+            auto value = *m_value;
+            mvApp::GetApp()->getCallbackRegistry().submitCallback([=]() {
+                mvApp::GetApp()->getCallbackRegistry().addCallback(getCallback(false), m_uuid, ToPyFloat(value), m_user_data);
+                });
+        }
 
     }
 
@@ -146,7 +151,12 @@ namespace Marvel {
         if (!m_enabled) m_disabled_value = *m_value;
 
         if (ImGui::DragInt(m_label.c_str(), m_enabled ? m_value.get() : &m_disabled_value, m_speed, m_min, m_max, m_format.c_str(), m_flags))
-            mvApp::GetApp()->getCallbackRegistry().addCallback(getCallback(false), m_uuid, nullptr, m_user_data);
+        {
+            auto value = *m_value;
+            mvApp::GetApp()->getCallbackRegistry().submitCallback([=]() {
+                mvApp::GetApp()->getCallbackRegistry().addCallback(getCallback(false), m_uuid, ToPyInt(value), m_user_data);
+                });
+        }
     }
 
     void mvDragFloat::handleSpecificKeywordArgs(PyObject* dict)